requirements.txt
Flask==2.3.3
Pillow==9.5.0
tensorflow==2.20.0
scikit-learn==1.5.0
numpy>=1.24.0
pandas==2.2.3
requests==2.31.0
Werkzeug==2.3.7
gunicorn==20.1.0
